Backened/requirements.txt
fastapi
uvicorn
openai-whisper
torch
python-multipart
email-validator
python-dotenv
groq